7 Добавить слайс для юзера
