+++ packages/client/src/app/store.ts
@@ -1,5 +1,6 @@
 import { configureStore } from '@reduxjs/toolkit'
-import authReducer, { Auth } from '@processes/auth/model/authSlice'
+import userReducer, { Auth } from '@processes/user/model/userSlice'
+//import authReducer, { Auth } from '@processes/auth/model/authSlice'
 
 export interface Store {
   auth: Auth
@@ -9,7 +10,7 @@ export type AppDispatch = typeof store.dispatch
 
 const store = configureStore({
   reducer: {
-    auth: authReducer,
+    auth: userReducer,
   },
 })
