+++ packages/client/src/pages/ProfilePage/ui/ProfilePassword.tsx
@@ -2,7 +2,7 @@ import { FC } from 'react'
 import s from './ProfilePage.module.scss'
 import { Button } from '@shared/components/ui/button'
 import { Form } from '@shared/components/ui/form'
-import { changePassword } from '@processes/profile/api/profileApi'
+import { changePassword } from '@processes/user/api/userApi'
 import { z } from 'zod'
 import { validationRules } from '@shared/config/validationRules'
 import { useForm } from 'react-hook-form'
