+++ packages/client/src/processes/user/api/userApi.ts
@@ -1,8 +1,26 @@
+import { BASE_AUTH_API } from '@processes/auth/api/authApi'
+
 const BASE_PROFILE_API = `${import.meta.env.VITE_API_URL}/user`
 
 export interface Password {
   oldPassword: string
   newPassword: string
+}
+
+export const user = async () => {
+  const response = await fetch(`${BASE_AUTH_API}/user`, {
+    method: 'GET',
+    headers: { 'Content-Type': 'application/json' },
+    credentials: 'include',
+  })
+
+  const data = await response.json()
+
+  if (!response.ok) {
+    throw new Error(data.reason)
+  }
+
+  return data
 }
 
 export const changePassword = async (credentials: Password) => {
